+++ src/models/transaction_dto.rs
@@ -2,7 +2,7 @@ use serde_derive::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct TransactionDTO {
-    pub valor: i32,
+    pub valor: i64,
     pub tipo: String,
     pub descricao: String,
 }
