+++ src/models/transaction_result.rs
@@ -2,6 +2,6 @@ use serde_derive::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct TransactionResult {
-    pub limite: i32,
-    pub saldo: i32,
+    pub limite: i64,
+    pub saldo: i64,
 }
